requirements.txt
Flask==3.0.3
flask-cors==4.0.0
gunicorn==21.2.0
Pillow==10.3.0
numpy==1.26.4
matplotlib==3.8.4